feat(front): add POC for ifc viewer using thatopen/engine_component
