+++ frontend_app_file/src/component/IfcViewer/IfcViewer.styl
@@ -1,0 +1,2 @@
+.ifcViewer
+  height 100%
